pyproject.toml
[project]
name = "linkloom"
version = "0.1.0"
description = "Lightweight, self-hosted bookmark manager."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler==3.11.0",
    "beautifulsoup4==4.13.4",
    "flask==3.1.1",
    "flask-login==0.6.3",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "httpx==0.28.1",
    "lxml==5.4.0",
    "pytest==8.4.1",
    "python-dateutil==2.9.0.post0",
    "rapidfuzz==3.14.1",
    "trafilatura==1.12.2",
]
